Enqueue touch events if found while expecting command answer
At any time, a touch event may be sent by the Nextion, even in the
middle of writing a command to the device. If an event is enqueued
before the command response, the command will appear to have failed,
even if successful, and additionally the event will be lost. Fix by
examining the header byte and implementing an event queue to stash
the event data until the next nexLoop() invocation, at which they
will be dispatched before reading more bytes from the serial port.