Skip all serial port initialization if baudrate is passed as 0
The default nexSerial initialization assumes that the Nextion device
starts with a speed of 9600 baud, to be changed at initialization
time, and also assumes that whichever serial port is to be used has
been programmed with the default pins for the platform. If either of
these assumptions is false, the initialization will reset to defaults
not adequate for the application. Fix by skipping the initialization
and baud change if the baud rate is explicitly 0. The speed initialization
is then left up to the calling code.